adding new structure in place to handle p2p and multiple communication channels
